Added patches and modified build to use CMake to build MSToolkit, Comet, and Kojak
