+++ ext/patches/MSToolkit/include/MSReader.h
